requirements.txt
chainlit==1.0.0
python-dotenv==1.0.0
metagpt 